# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  alioth-hrc_internal_model.Ledger:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  alioth-hrc_internal_service.ContactStatRow:
    properties:
      balance_cents:
        type: integer
      contact_id:
        type: integer
      last_occurred_on:
        type: string
      total_give_cents:
        type: integer
      total_receive_cents:
        type: integer
    type: object
...
  alioth-hrc_internal_service.LedgerCreate:
    properties:
      name:
        maxLength: 200
        type: string
    required:
    - name
    type: object
  alioth-hrc_internal_service.LedgerUpdate:
    properties:
      name:
        maxLength: 200
        type: string
    required:
    - name
    type: object
...
  alioth-hrc_internal_service.MonthlyStatRow:
    properties:
      give_cents:
        type: integer
      receive_cents:
        type: integer
      year_month:
        type: string
    type: object
...
  alioth-hrc_internal_service.SummaryResult:
    properties:
      balance_cents:
        type: integer
      ledger_id:
        type: integer
      monthly:
        items:
          $ref: '#/definitions/alioth-hrc_internal_service.MonthlyStatRow'
        type: array
      total_give_cents:
        type: integer
      total_receive_cents:
        type: integer
      year:
        type: integer
    type: object
...
  internal_handler.ErrorJSON:
    properties:
      error:
        type: string
    type: object
...
  internal_handler.LedgerListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/alioth-hrc_internal_model.Ledger'
        type: array
    type: object
  internal_handler.OKStatus:
    properties:
      status:
        type: string
    type: object
...
  internal_handler.StatsContactsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/alioth-hrc_internal_service.ContactStatRow'
        type: array
    type: object
info:
  contact: {}
  description: 人情往来记账后端：认证、联系人、账本、礼金分类与流水、统计等。
  title: alioth-hrc API
  version: "1.0"
paths:
...
  /ledgers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.LedgerListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 账本列表
      tags:
      - ledgers
    post:
      consumes:
      - application/json
      parameters:
      - description: 账本
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/alioth-hrc_internal_service.LedgerCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/alioth-hrc_internal_model.Ledger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 创建账本
      tags:
      - ledgers
  /ledgers/{id}:
    delete:
      parameters:
      - description: 账本 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.OKStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 删除账本
      tags:
      - ledgers
    get:
      parameters:
      - description: 账本 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alioth-hrc_internal_model.Ledger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 账本详情
      tags:
      - ledgers
    put:
      consumes:
      - application/json
      parameters:
      - description: 账本 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 账本
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/alioth-hrc_internal_service.LedgerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alioth-hrc_internal_model.Ledger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 更新账本
      tags:
      - ledgers
...
  /stats/contacts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.StatsContactsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 按联系人统计
      tags:
      - stats
  /stats/summary:
    get:
      parameters:
      - description: 账本 ID
        in: query
        name: ledger_id
        type: integer
      - description: 筛选年份（不传则按全部时间）
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alioth-hrc_internal_service.SummaryResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorJSON'
      security:
      - BearerAuth: []
      summary: 汇总统计
      tags:
      - stats
securityDefinitions:
  BearerAuth:
    description: 访问令牌，格式：Bearer {access_token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"